+++ algorithms/__init__.py
@@ -1,2 +1,3 @@
 # flake8: noqa
 from algorithms.sorting.bubble_sort import BubbleSort
+from algorithms.sequences.fibonnaci_sequence import FibonacciSequence